+++ runtime-kotlin/runtime-core/src/test/kotlin/skillbill/architecture/RuntimeArchitectureTest.kt
@@ -1931,6 +1931,7 @@ class RuntimeArchitectureTest {
       "skillbill.workflow.taskruntime.model.FeatureTaskRuntimeResolvedBranch.fromArtifactMap",
       "skillbill.workflow.taskruntime.model.FeatureTaskRuntimeGoalContinuationArtifact.toArtifactMap",
       "skillbill.workflow.taskruntime.model.FeatureTaskRuntimeGoalContinuationArtifact.fromArtifactMap",
+      "skillbill.workflow.taskruntime.model.FeatureTaskRuntimeGoalPlanningImport.toArtifactMap",
       "skillbill.workflow.taskruntime.model.GoalSubtaskReviewCompactFinding.toArtifactMap",
       "skillbill.workflow.taskruntime.model.GoalSubtaskReviewCompactFinding.fromArtifactMap",
       "skillbill.workflow.taskruntime.model.GoalSubtaskReviewPassResult.toArtifactMap",